requirements_cpu.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fsspec==2025.3.2
ftfy==6.3.1
greenlet==3.2.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
intel-openmp==2021.4.0
jinja2==3.1.6
jose==1.0.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
open-clip-torch==2.32.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
portalocker==2.10.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pyyaml==6.0.2
qdrant-client==1.13.3
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
safetensors==0.5.3
setuptools==79.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.2
sympy==1.13.1
tbb==2021.11.0
timm==1.0.15
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.67.1
typer==0.15.2
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
minio==7.2.15
geopy==2.4.1